+++ R/is_logical.R
@@ -3,8 +3,15 @@
 #' Check that `x` is a length-one logical vector with only allowed values.
 #'
 #' @param x object to test.
-#' @param allow_zero_length `TRUE` or `FALSE`: allow zero-length `x` of the correct type?
-#' @param allow_NA `TRUE` or `FALSE`: allow `NA`s of the correct type in `x`?
+#' @param allow_zerolength `TRUE` or `FALSE`: allow
+#' [zero-length][is_zerolength()] `x` of the correct type?
+#' @param allow_NA `TRUE` or `FALSE`: allow [NA]s of the correct type in `x`?
+#'
+#' @details
+#' `is_logical()` returns `TRUE` for `x` with length one, and for logical-type
+#' [zero-length][is_zerolength()] `x` if `allow_zerolength` is `TRUE`.
+#'
+#' `is_logical()` returns `TRUE` for (logical-type) `NA` if `allow_NA` is `TRUE`.
 #'
 #' @returns `TRUE` or `FALSE` indicating if `x` is a length-one logical vector
 #' only containing allowed values.
@@ -13,7 +20,7 @@
 #' collections of checks on type and length
 #'
 #' @seealso
-#' The vignette *Design choices regarding function input*:
+#' The vignette *Design choices*:
 #' `vignette("design_choices", package = "checkinput")`.
 #'
 #' @examples
@@ -25,10 +32,12 @@
 #' is_logical(NA_character_, allow_NA = TRUE) # FALSE: incorrect type
 #'
 #' @export
-is_logical <- function(x, allow_zero_length = FALSE, allow_NA = FALSE) {
-  stopifnot(is.logical(allow_zero_length), length(allow_zero_length) == 1L, !is.na(allow_zero_length),
-            is.logical(allow_NA), length(allow_NA) == 1L, !is.na(allow_NA))
+is_logical <- function(x, allow_zerolength = FALSE, allow_NA = FALSE) {
+  stopifnot(is.logical(allow_zerolength), length(allow_zerolength) == 1L,
+            !is.na(allow_zerolength), is.logical(allow_NA),
+            length(allow_NA) == 1L, !is.na(allow_NA))
+
   is.logical(x) && is.atomic(x) && is.null(dim(x)) &&
-    (length(x) == 1L || (allow_zero_length && length(x) == 0L)) &&
+    (length(x) == 1L || (allow_zerolength && length(x) == 0L)) &&
     (allow_NA || !anyNA(x))
 }