+++ inst/tinytest/test_is_character.R
@@ -14,10 +14,10 @@ for(x in list(c("a", "b"), 1, NA_character_)) {
 
 #### Tests ####
 for(allow_empty in false_true) {
-  for(allow_zero_length in false_true) {
+  for(allow_zerolength in false_true) {
     for(allow_NA in false_true) {
       expect_silent(expect_true(is_character(x = "nco", allow_empty = allow_empty,
-                               allow_zero_length = allow_zero_length, allow_NA = allow_NA)))
+                               allow_zerolength = allow_zerolength, allow_NA = allow_NA)))
     }
   }
 }
@@ -27,50 +27,50 @@ for(x in list(NULL, FALSE, TRUE, NA, NaN, numeric(0), NA_real_, 0, 314, Inf,
               c("nco", "ibf"), data.frame(a = "nco"), matrix("nco"),
               list("nco"))) {
   for(allow_empty in false_true) {
-    for(allow_zero_length in false_true) {
+    for(allow_zerolength in false_true) {
       for(allow_NA in false_true) {
         expect_silent(expect_false(is_character(x = x, allow_empty = allow_empty,
-                                  allow_zero_length = allow_zero_length, allow_NA = allow_NA)))
+                                  allow_zerolength = allow_zerolength, allow_NA = allow_NA)))
       }
     }
   }
 }
 
 # x for which return of is_character() is equal to argument 'allow_empty'
-for(allow_zero_length in false_true) {
+for(allow_zerolength in false_true) {
   for(allow_NA in false_true) {
     expect_silent(expect_true(is_character(x = "", allow_empty = TRUE,
-                             allow_zero_length = allow_zero_length, allow_NA = allow_NA)))
+                             allow_zerolength = allow_zerolength, allow_NA = allow_NA)))
     expect_silent(expect_false(is_character(x = "", allow_empty = FALSE,
-                              allow_zero_length = allow_zero_length, allow_NA = allow_NA)))
+                              allow_zerolength = allow_zerolength, allow_NA = allow_NA)))
   }
 }
 
-# x for which return of is_character() is equal to argument 'allow_zero_length'
+# x for which return of is_character() is equal to argument 'allow_zerolength'
 for(allow_empty in false_true) {
   for(allow_NA in false_true) {
     expect_silent(expect_true(is_character(x = character(0), allow_empty = allow_empty,
-                             allow_zero_length = TRUE, allow_NA = allow_NA)))
+                             allow_zerolength = TRUE, allow_NA = allow_NA)))
     expect_silent(expect_false(is_character(x = character(0), allow_empty = allow_empty,
-                              allow_zero_length = FALSE, allow_NA = allow_NA)))
+                              allow_zerolength = FALSE, allow_NA = allow_NA)))
   }
 }
 
 # x for which return of is_character() is equal to argument 'allow_NA'
 for(allow_empty in false_true) {
-  for(allow_zero_length in false_true) {
+  for(allow_zerolength in false_true) {
     expect_silent(expect_true(is_character(x = NA_character_, allow_empty = allow_empty,
-                             allow_zero_length = allow_zero_length, allow_NA = TRUE)))
+                             allow_zerolength = allow_zerolength, allow_NA = TRUE)))
     expect_silent(expect_false(is_character(x = NA_character_, allow_empty = allow_empty,
-                              allow_zero_length = allow_zero_length, allow_NA = FALSE)))
+                              allow_zerolength = allow_zerolength, allow_NA = FALSE)))
   }
 }
 
 # Arguments that should result in an error.
-for(allow_zero_length in false_true) {
+for(allow_zerolength in false_true) {
   for(allow_NA in false_true) {
     expect_error(is_character(x = "nco", allow_empty = NA,
-                              allow_zero_length = allow_zero_length, allow_NA = allow_NA),
+                              allow_zerolength = allow_zerolength, allow_NA = allow_NA),
                  pattern = "is_logical(allow_empty) is not TRUE", fixed = TRUE)
   }
 }
@@ -78,19 +78,19 @@ for(allow_zero_length in false_true) {
 for(allow_empty in false_true) {
   for(allow_NA in false_true) {
     expect_error(all_characters(x = "nco", allow_empty = allow_empty,
-                                allow_zero_length = NA, allow_NA = allow_NA),
-                 pattern = "is_logical(allow_zero_length) is not TRUE", fixed = TRUE)
+                                allow_zerolength = NA, allow_NA = allow_NA),
+                 pattern = "is_logical(allow_zerolength) is not TRUE", fixed = TRUE)
   }
 }
 
 for(allow_empty in false_true) {
-  for(allow_zero_length in false_true) {
+  for(allow_zerolength in false_true) {
     expect_error(all_characters(x = "nco", allow_empty = allow_empty,
-                                allow_zero_length = allow_zero_length, allow_NA = NA),
+                                allow_zerolength = allow_zerolength, allow_NA = NA),
                  pattern = "is_logical(allow_NA) is not TRUE", fixed = TRUE)
   }
 }
 
 
 #### Remove objects used in tests ####
-rm(allow_empty, allow_NA, allow_zero_length, false_true, x)
+rm(allow_empty, allow_NA, allow_zerolength, false_true, x)